Add x64 build configuration and change rendering API to DirectX 9

+++ Code/Libraries/Source/Compression/LZHCompress/CompLibSource/LZHLCompressor.cpp
@@ -56,7 +56,7 @@ size_t LZHLCompressor::compress( BYTE* dst, const BYTE* src, size_t sz ) {
   }
 
   for (;;) {
-    int srcLeft = srcEnd - src;
+    int srcLeft = (int)(srcEnd - src);
     if ( srcLeft < LZMATCH ) {
       if ( srcLeft ) {
         _toBuf( src, srcLeft );
